Fixed deleteCard working but not removing the card from the UI. Async leaves a delay. Solution: useState to hide the card while async works.
